requirements.txt
flask-genshi>=0.4
flask-script
ordereddict
pyyaml
simplejson
pystemmer
pqs
Twisted
jellyfish
